requirements.txt
fastapi 
uvicorn 
torch 
transformers 
peft 
bitsandbytes 
trl 
datasets
scipy
python-multipart
boto3
python-dotenv